everywhere: Avoid shared_ptr<promise>
The design of future and promise is such that they can track each
other as they move. There is no need to have a stable heap allocation.

More importantly, some of the code using shared_ptr was fairly
hard to understand. It looked like

auto pr = make_lw_shared(promise<>());
...
}).then_wrapped([pr] (future<> f) {
    f.forward_to(std::move(*pr));
));
...
return pr->get_future();

Note that the forward_to can be executed before get_future. The reason
this works is that forward_to is defined as

void forward_to(promise<T...>&& pr) noexcept {
    if (state()->available()) {
        state()->forward_to(pr);
    } else {
        _promise->_future = nullptr;
        *_promise = std::move(pr);
        _promise = nullptr;
    }
}

If the value is available, which it is for the future passed to
then_wrapped, forward_to doesn't actually move anything.

With this patch we avoid depending on when forward_to will or will not
move its argument.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>