test_runner::start: wait until run_deprecated either starts running or returns an error.
Fixes a crash when running the test with -- --help, or an invalid option
e.g. -- --foo.

Note that when an error is returned by run_deprecated, e.g. on --help
or unrecognized option, start() gets the error via the _ready promise
and exits with the error exit code, rather than returning an
initialization error - to avoid the annoying boost message in this case.

Fixes #673

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>