seastar: add abort() function
First, flush any buffered log output to the standard output
and then call ::abort() that will print the backtrace message.

The main reason for doing that is not losing log messages
when aborting the program.  Also, it will improve the correlation
of the backtrace to other events printed to the log.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>