file_impl: add defer_close
Keep track of the file_impl _closed state and return it with is_closed().
Allow implementation to mark_as_closed().
And provide defer_close() to ask the destructor to close the file if open upon destruction.

Regardless, the caller should explicitly call close() and the deferred close
is implemented just as a safety precaution.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>