rpc: unregister_handler: move handler to _unregistered_handlers
And do not destroy it since in-flight message may still be using it
as seen in https://github.com/scylladb/scylla/issues/5228

Use lw_shared_ptr<rpc_handler> for keeping track of the
underlying rpc_handler instance and moving it from `_handlers` to
`_unregistered_handlers` upon unregistration.

get_handler still returns a pure pointer that is supposed to be valid
throught the life cycle of call.

Note that we cannot return a shared ptr in get_handler and use
that to keep the handler around if unregistered while processing
since get_handler might be called by a different cpu than the
one used to construct the rpc_handler - hence the need
to access the rpc_handler using a pure pointer to a global instance.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>