rpc: let unregister_handler return a future<>
The future is set with a value when the shared rpc_handler
is released.

This extends the api.  Previously unregister_handler returned void,
so callers will now be warned of discarding the returned future
if they don't wait on.  Callers should consider doing that since
it may be unsafe to proceed after unregister_handler when there
could be asynchronous users using the handler.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>