log-cli: add string-based extract_settings function
So that the program can also pass string values for extracting
logging settings rather than program_options::map.

This way, if the program gets the options from anywhere else
but the command line options (e.g. yaml config file), it doesn't
have to understand the type and semantics of the options and be able to
parse them.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>