reactor: introduce same_file
The function checks if two paths refer to the same file on the filesystem,
as hard links can be linked to the same inode.

This could be used for example to make `link_file` idempotent
by detecting if the destination link already exists
that it refers to the same file as the source.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>