utils: link_file: add allow_same flag
allow_same determines whether the function succeeds if
oldpath and newpath are already linked to the same file.

allow_same::yes allows link_file to be idempotent
so it can be used when retrying the link operation.

allow_same::no provides the backward compatible semantics
of the link(3P) posix system call.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>