tests: fstream: do not abandon futures
No need to run stuff in background and to synchronize using a semaphore.

Fixes:
*** 2 abandoned failed future(s) detected

Historical notes:
The original test that was committed in 10a71dada by Raphael didn't have the semaphore.
The latter was added in 83998a893 by Shlomi as "Boostify the fstream_test"
so it's use is not part of the test itself, but rather related to the test environment.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>