file: futurize make_file_impl
Specified as noexcept as fstat and fstatfs are now noexcept.
Any failure to make_shared a file_impl will be caught
in a continuation functor and will be returned as an exception
future.

Use make_file_impl in open_file_dma and open_directory.
With that, the following constructor can go away:
  explicit file(int fd, file_open_options options)

Next steps are to make open_file_dma and open_directory noexcept
and to specify class file constructors as noexcept.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>