smp_message_queue: submit: make noexcept
So we can make future-returning functions depending on it
noexcept too.

Failure to allocate in infrastructure-level functions like this one
will crash the system, until we handle it, e.g. by dipping into
reserved memory.  To reflect that, disable_failure_guard was added
so that allocation error injection in testing will be disabled here.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>