gate: make with_gate noexcept
Define gate::do_with that returns an exception
future rather than throwing if the gate is closed.

Call gate::do_with from with_gate in SEASTAR_API_LEVEL 3.

Note that this is an API change. If someone is trying to
catch a gate_closed_exception they will now have to look
at the future.  This behavior is kept in SEASTAR_API_LEVEL 2.

Modify the test_with_gate unit test to test both API versions;
but don't bump the API version just yet, this will be done
in the next patch.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>