gate: make with_gate noexcept
Currently, with_gate returns a future<> that may hold an exception
as returned by the called `func`.  However, if the gate is closed,
with_gate throws a gate_closed_exception, similar to gate::enter()
or gate::check().

This subtlety has proved to be error prone since callers of with_gate,
are either not aware that it may throw and fail to handle the exception
or expect it to return the error as an exceptional future as most other
future-returning functions do, by convention.  See example below
of existing code and how it should be fixed following this patch.

This patch defines gate::with_gate that returns an exception
future rather than throwing if the gate is closed.

Note that this is an API change. If someone is trying to
catch a gate_closed_exception they will now have to look
at the future.  This behavior is kept in SEASTAR_API_LEVEL 2.

Call gate::with_gate from with_gate in SEASTAR_API_LEVEL 3.

Add test_with_gate unit test to test both API versions;
but don't bump the API version just yet, this will be done
in the next patch.

For example, current code that is currently written in this way:
    try {
        return with_gate(g, [] {
            /* do stuff */
        }).handle_exception([] (std::exception_ptr) {
            /* handle errors */
        });
    } catch (gate_closed_exception& e) {
        /* handle gate_closed case */
    }

Should be converted to:
    return with_gate(g, [] {
        /* do stuff */
    }).handle_exception_type([] (gate_closed_exception& e) {
        /* handle gate_closed case */
    }).handle_exception([] (std::exception_ptr) {
        /* handle errors */
    });

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>