api-level: Add version 3
This adds api version 3 and expands the infrastructure a bit for when
time comes to add a v4.

The idea is that only one of the SEASTAR_INCLUDE_API_V* macros is
inline. That alone takes care of exposing the APIs introduced in a
version, but we also have to bring forward APIs introduced in previous
versions. That is done with using directives.

To avoid having to have a using directive for each function or type,
the api_vX namespace is divided into more namespaces. Right now we are
just adding v3, so all additions are in a single api_v3::and_newer
namespace.

When a v4 is added, it can just have "using api_v3::and_newer". If v4
also modifies an api in v3, that api can be moved from and_newer
directly to v3 and it will not be in v4.

The strategy was tested while there was a v1, so we know it works well
with 3 versions.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>