lock: specify with_lock as noexcept
The read and write lock paths can be safely specified as noexcept
as the semaphore primitives they use are noexcept.

Specify semaphore::try_wait as noexcept (it trivially is).

The higher level `with_lock` function made noexcept using
futurize_invoke of its function body's implementation
that is noexcept iff the func passed to it is nothrow_move_constructible.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>