lock: specify with_lock as noexcept
The read and write lock paths can be safely specified as noexcept
as the semaphore primitives they use are noexcept.

Specify semaphore::try_wait as noexcept (it trivially is).

Note that the unlock path may throw from semaphore::signal() -> promise::set_value(),
but with_lock may still be specified as noexcept as unlock is invoked
in then_wrapped continuation.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>