future: add with_nested_exception
Returns a failed future optionally with a nested exception
of another future if the other one failed as well.

Required for the following patch that closed a file
on failure and returns the original future in a then_wrapped
clause.  It can't use `finally` since the latter erases
the inner future's type and returns the outer future's type
and here we need the opposite - maybe return the outer's
future (f.close()) exception as a nested exception of the
original future's exception.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>