everywhere: Work around gcc bug 95368
With this it is possible to build seastar with an assert added to
require continuations to be nothrow move constructible.

This patch intentionally does nothing more than use the
SEASTAR_GCC_BZ95368 macro when needed. There can still be bugs with
the initial lambda construction throwing, but they were already present
and are better fixed in a followup patch series.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>