Add a new API level without seastar::subscription
The subscription we have in seastar was just a wrapper over a
future<>. This patch adds a new API version that directly uses
future<> instead of subscription.

This makes the existing APIs simpler, and, once the old API is
selected, the name seastar::subscription can be used for something
else.

I have tested scylla with both the current and, with a patch, the new
API.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>