util: Add a disable_failure_guard to deferred_action
It is not safe to use defer when we want to be exception safe, since
an exception cannot propagate out of the destructor. That is a know
restriction, but we still want to use it in contexts where calling
std::unexpected on std::bad_alloc is reasonable (tests for example).

Found while looking at https://github.com/scylladb/scylla/issues/6831

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>