Always allocate in repeat
This makes it possible to use only repeat in cases that now use
repeat+do_with.

On its own this patch is probably too expensive. What is probably
needed is a repeat_with variant. With that we can replace the uses of
repeat+do_with and then see if we can make repeat==repeat_with.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>