defer: add run_now method
For running the deferred action once, on-demand.

This simplifies a rather common use case (e.g. in scylla tests)
where a resource needs to be destroyed both implicitly, when the
deferred action goes out of scope (either normally on the exception
path), and also explicitly, in the middle of the code block.

Then, the following:
{
    auto obj1 = make_object1();
    deferred_action auto_close = defer([&obj1] { obj1.close(); });

    // do something
    auto obj2 = make_object2();

    auto_close.cancel();
    obj1.close();

    // do something with obj2
}

becomes:
{
    auto obj1 = make_object1();
    deferred_action auto_close = defer([&obj1] { obj1.close(); });

    // do something
    auto obj2 = make_object2();

    auto_close.run_now();

    // do something with obj2
}

Without `run_now`, we can't simply enclose the use of obj1 in
a block to ensure that auto_close will be invoked when it goes out of
scope since obj2 is made inside this block and is needed after
obj1 is closed.

Explcitly canceling the deferred action and running the action
is undesired since it leads to code duplication when the action
is implemented twice: once as a deferred action, and then as
an explicit action.

Test: defer_test(dev)
Signed-off-by: Benny Halevy <bhalevy@scylladb.com>