stall-analyser: track total/count in Link not Node
Track the latency statistics in the Graph's Links,
not in the Nodes, since the percentages printed
are relative to the siblings caller/callee node,
not absolute.

When a node (address) calls multiple other nodes,
or being called by multiple other nodes, each link
is printed with its total/count statistics.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>