Add deferred_close and deferred_stop
Close/stop the object when the respected
deferred_{close,stop} object is destroyed.

In addition, the deferred close/stop classes
provide {close,stop}_now() methods that
can be used to perform the cleanup once,
while the deferred_{close,stop} is still alive.

Also, add a unit test to demonstrate and test the interface.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>