rwlock: add close method and lock_close_exception
Allow closing a rwlock so that every consecutive attempt
to acquire the lock will fail with `lock_closed_exception`
similar to closed gates and broken semaphore.

The motivation is for using the rwlock for shared/exclusive
access to a resource and preveting any further access before
destroying the said resource by closing it.

For example, we can apply such closeable rwlock to synchronize
async access to scylla's table, instead of using a gate.

Also, added respective unit test.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>