fstream: expose io intent via file_input_stream_options
Allow the file opener to hold on to an io_intent
and pass an optional pointer to it when making the file input stream
using the file_input_stream_options.

The io_intent can be used later on to cancel outstanding
io requests, e.g. when aborting a reader due to an
external timeout.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>