future: loop functions: standardize preemption and error handling
Optimize for the common, error-free case.
If need_preempt, let the the state-allocated path handle
both success and error cases, otherwise, loop inside a try/catch
block and let fut.get()/get0() throw if the future contains an
error, and then handle it in the catch clause, so to tighten
the inner loop around the error-free path.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>