reactor: destroy_scheduling_group: make sure scheduling_group is valid
When called with an invalid sg (e.g. a previously destroyed one)
with_scheduling_group segfaults when dereferencing _task_queues[sg._id]
which is nullified by destroy_scheduling_group.

To prevent that, verify first the the sg is valid, and
abort after calling on_internal_error_noexcept otherwise.
return an exceptional future in debug mode to enable a unit test.

While at it, also mark the method noexcept.

Added a debug unit tests to verify the returned exception
when re-destorying a scheduling_group or when destroying an
invalid scheduling_group.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>