queue: mark methods depending on std::queue noexcept
Although the respective std::queue methods are not
specified as noexcept they have no reason to throw.

To simplify marking of methods that rely on those as noexcept,
mark the low-level methods as noexcept.

Note that seastar::queue::pop() is mark
noexcept(std::is_nothrow_move_constructible_v<T>)
since it moves the front data element out of the queue
and returns it.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>