queue: pop_eventually: do not duplicate not_empty logic and mark noexcept
not_empty does already returns an exceptional_future
if one is set and returns a ready future if the
queue is !empty(), so no reason to repeat that
in pop_eventually.

With that it can be marked noexcept.

Note that a check for exception was kept when not_empty
resolves to preserve current semantics.
Ideally, pop() would just throw if an exception is set
but consume() relies on it to not throw in this case
so it can drain the queue.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>