file: append_challenged_posix_file: close: harden against failures
Close may return an error if failed to flush pending operations,
but it should always succeed in the sense that the file can be safely
destroyed even when an error is returned.

Make sure it does a best effort to truncate the file,
and that it retuns any error during the close process.

In all cases, it calls posix_file_impl::close()
and sets _closing_state = state::closed so that
the destructor can test that and allow logical vs.
commited size discrepencies.

Add a unit test that reproduces the failed assertion
without this fix and tests the close error path
using alloc failure injection.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>