shared_mutex: make with_* function noexcept
Define variants of with_shared and with_lock
for template parameter functions that may throw
when moved.

In this case, they need to be captured by value
before the lock is acquired to ensure
the error is returned as an exceptional future
and that the lock is released properly when taken.

This patch fixes a bug where currently,
if the function throws when moved,
the shared_mutex remains locked.

Add unit test that reproduces the issue without
this patch.

Fixes #970

Test: test_shared_mutex_throwing_func (w/ clang++ c++17/c++20)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>