tcp: tcb: close: handle errors
wait_for_all_data_acked might return an error
set to _all_data_acked_promise.

_all_data_acked_promise is set with tcp_reset_error()
by do_reset() so handle it as an expected error.

Print a warning and swallaw all other errors
while setting the _state directly to CLOSED
as close() must always succeed.

With that, it can be marked as noexcept.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>