semaphore: add abortable get_units
Provide a get_unit function that accepts an abort_source
allowing the caller to abort get_units from an external source.

For example, the intention is to use it to in scylla
when compaction_manager tasks are stopped while waitin
for a semaphore.

Also, a later patch will make use of this for acquiring
rwlock locks.

Add a respective unit test.

Test: semaphore_test

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>