future: loop: add do_for_each_until
Modeled after `repeat` and `repeat_until_value`,
`do_for_each_until` calls a function on each of the
items in a range until the range is exhausted,
the function return stop_iteration::yes, or the function fails.

Although this functionality can be achieved without
this specialized template, it will require keeping an
external state (using e.g. do_with) to do nothing
in the function after the stop point is reached,
since the function is called on all items unconditionally,
or by returning a particular exception for the permature
stopping of the do_for_each loop, and handling it.

Both options are sub-optimal.

Like do_for_each, do_for_each_until optimizes the synchronous
case where the function keeps returning a ready future,
and it delayes its allocation until the function returns
an unavailable future or when preemption is needed.

Test: unit(release)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>