coroutine: exception: log_exception_trace when making an exception
Like we do in `make_exception_future(Exception&& ex)`,
we want a trace of where the exception originated, but in these cases it
is never thrown, but rather an exception object is created
and is propagated up the stack via exceptional futures.
(We use _Unwind_RaiseException to trace "real" exceptions).

Fixes #1091

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>