coroutine: add parallel_for_each_check_preemption
Add a variant of parallel_for_each that checks
if preemption is needed to prevent stalls caused
by a large number of elements we're iterating over.

Note that even if the called function returns a non-ready
future, we currently just store it in a vector and
continue on to the next eleement, without preempting.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>