semaphore: assert no outstanding units when moved
semaphore_units keeps a pointer to the semaphore
soe if the semaphore is moved and/or destroyed
destroying the sempahore_units will caused
use-after-move or use-after-free.

This change adds _outstanding_units counter member
to basic_semaphore and manages it in semaphore_units.
semaphore move constructor then asserts that
o._outstanding_units == 0 when moved.

Add a unit test to semaphore_test to cover this case.

Refs #1323

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>