gate: add gate holders debugging
gate::holder keeps a naked pointer to the gate it holds.
Its life time is independent of the gate and there is
no mechanism to update this pointer if the gate is moved,
therefore if the gate is moved or destroyed while there are
outstanding holders pointing to it there is a risk for
use-after-move or use-after-free.

Count the number of gate holders in debug mode
and assert it's zero when the gate is moved or destroyed.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>