condition_variable: wait: simplify waiter allocation
set_callback() and arm() are both noexcept so there's
no need to cleanup the timeout_waiter allocation
for them.

std:;bind might still through but:
1. failing to create this function is equivalent
to failing to allocate a continuation, for which
we already terminate by convention.
2. The wait() function itself is already marked
noexcept, so recovering from this allocation failure
is futile in any case.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>