semaphore: semaphore_units: return units when reassigned
When semaphore_units are (move-) reassigned with
other units, the held units aren't currently
returned to the semaphore.

This change implements the move-reassignment operator
by swapping the semaphore and the units counter
so that when the moved-from semaphore_units object
is destroyed, the units will be returned to the semaphore
via `~semaphore_units` -> `return_all`.

Add a unit test reproducer.

Fixes #1465

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>