semaphore: disallow move after used
Referring to #1323 and discussion in #1477.

Avi Kivity <avi@scylladb.com> asserted that:
> We have a choice:
> declare that most things can be movable and live with the complication and slowdown
> declare that control structures (like semaphores) have stable addresses (as they typically do anyway)
> There's really no reason to move a sempahore.

Gleb Natapov <gleb@scylladb.com> proposed:
> add an assertion in debug mode if a semaphore is moved after the first use. It can be a simple bool that is set to true on the first wait().

Since we still need to be able to move the semaphore
after it created this patch adds `bool _used` and implements
the move ctor, asserting that the semaphore is moved before used.

Close #1323

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>