future: future_state_base: assert owner shard in debug mode
Assert that the future state is accessed only
on the shard on which it was created as the future
and promise objects are not designed to be accessed
across shards.

For example, currenty, if promise.set_value() (or
set_exception) is called on a promise from a foreign
shard the promise.get_future() continuation would
run on the foreign shard.

Fixes scylladb/seastar#1563

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>