future: future_state_base: implement the move-assignment operator
Before adding assert_shard_owner in debug mode
in the next patch.

The move-assignment implementation is subtle as
we need to handle to self-move-assign case in
any::move_it.

Without handling the self-move case, futures_test
crashes on the exit path due to the task
test_run_in_background leaves behind in the background.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>